+++ libfreerdp/codec/rfx.c
@@ -162,7 +162,7 @@ static void rfx_tile_init(void* obj)
 
 static void* rfx_decoder_tile_new(const void* val)
 {
-	const size_t size = 4 * 64 * 64;
+	const size_t size = 4ull * 64ull * 64ull;
 	RFX_TILE* tile = NULL;
 	WINPR_UNUSED(val);
 
@@ -265,7 +265,7 @@ RFX_CONTEXT* rfx_context_new_ex(BOOL encoder, UINT32 ThreadingFlags)
 	 *
 	 * We then multiply by 3 to use a single, partioned buffer for all 3 channels.
 	 */
-	priv->BufferPool = BufferPool_New(TRUE, (8192 + 32) * 3, 16);
+	priv->BufferPool = BufferPool_New(TRUE, (8192ull + 32ull) * 3ull, 16);
 
 	if (!priv->BufferPool)
 		goto fail;
@@ -574,7 +574,7 @@ static BOOL rfx_process_message_channels(RFX_CONTEXT* context, wStream* s)
 	}
 
 	/* Now, only the first monitor can be used, therefore the other channels will be ignored. */
-	Stream_Seek(s, 5 * (numChannels - 1));
+	Stream_Seek(s, 5ull * (numChannels - 1));
 	WLog_Print(context->priv->log, WLOG_DEBUG,
 	           "numChannels %" PRIu8 " id %" PRIu8 ", %" PRIu16 "x%" PRIu16 ".", numChannels,
 	           channelId, context->width, context->height);
@@ -2259,7 +2259,7 @@ static INLINE BOOL rfx_write_progressive_region(RFX_CONTEXT* rfx, wStream* s,
 	 * RDPRFX:   LL3, LH3, HL3, HH3, LH2, HL2, HH2, LH1, HL1, HH1
 	 * RDPEGFX:  LL3, HL3, LH3, HH3, HL2, LH2, HH2, HL1, LH1, HH1
 	 */
-	for (UINT16 i = 0; i < msg->numQuant; i++)
+	for (size_t i = 0; i < msg->numQuant; i++)
 	{
 		const UINT32* qv = &msg->quantVals[i * 10];
 		/* RFX_COMPONENT_CODEC_QUANT */
